Turn most of my function into autoload func
Move functions from the ~/bin directory to zsh/zfunction which contains
autoload functions.
also made some variable local in order not to pollute the shell
namespace.
rem_save could not be moved as it needs to be in $PATH for remind to be
able to recognize it as a command.