Use git grep instead of the silver searcher
I decided to move from the silver searcher to git grep. If inside a git
project grepprg is set to git grep, otherwise the default setting is
kept (grep).
I needed a function that checks if I'am inside a git project and another
that set grepprg depending on the output of the first function.
I also added a autocommand group that is launch when entering vim that
call the latter function to set grepprg.