Add abcde configuration file
This program allows to rip audio cd.
It has been set to rip cd into mp3-cbr320, retrieve the cover and
follows xdg standard. The temporary wav filed are stored in /tmp instead
of the current directory in order not to stress my ssd.
Note: perl packages from aur have to be installed:
webservice-musicbrainz0 and musicbrainz-discid. For some reasons, it
didn't work with webservice-musicbrainz package.