Add cookiecutter templates for php projects
This commit is composed of two tightly related cookicutter templates:
- one that build a base php project based on docker for php + apach +
mysql + mailserver. The cookiecutter command is to be called from the
directory that will store the php project.
- one that create a tmux session scripts for this particular type of php
project that open the editor, open firefox, watch files to reload the
browser on changes and starts the docker stacks. This cookiecutter
command to build the session is to be called from the tmp directory.