Show/Remove non breakable spaces
Bépo keyboard layout sometimes (especially when working with css files
for some reason) makes me insert nbsp that can only be seen if 'list' is
on. I have added a nbsp indicator on the statusline and a command that
allow me to replace all nbsp with space in the whole buffer.